lfq.c: Use XCHG instead of CAS where appropriate.  Remove ctx->count updates.
The functions that append to the end of a linked list were using CAS retry loops
simply to get the old value of the tail pointer while modifying it.

That's what XCHG is for.

__atomic builtins would let us specify sequential-consistency semantics
instead of maybe having something weaker on non-x86.  But for now only
try to make this work on x86 and use a full-barrier exchange.